fix bug with cached linkage
